third_party/blink/web_tests/platform/generic/fast/css/parsing-object-position-expected.txt

This tests checks that all of the input values for object-position parse correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testComputedStyle(";") is "50% 50%"
PASS testComputedStyle("object-position: 10px;") is "10px 50%"
PASS testComputedStyle("object-position: 10px 10px;") is "10px 10px"
PASS testComputedStyle("object-position: right top;") is "100% 0%"
PASS testComputedStyle("object-position: top right;") is "100% 0%"
PASS testComputedStyle("object-position: left 20px bottom 50%;") is "20px 50%"
PASS testComputedStyle("object-position: right 50% bottom 25%;") is "50% 75%"
PASS testComputedStyle("object-position: bottom 20px right 12px;") is "calc(100% - 12px) calc(100% - 20px)"
PASS test("object-position: inherit;") is "inherit"
PASS test("object-position: initial;") is "initial"
PASS test("object-position: left;") is "left center"
PASS test("object-position: top;") is "center top"
PASS test("object-position: top right;") is "right top"
PASS test("object-position: right top;") is "right top"
PASS test("object-position: center center;") is "center center"
PASS test("object-position: center;") is "center center"
PASS test("object-position: bottom center;") is "center bottom"
PASS test("object-position: left center;") is "left center"
PASS test("object-position: bottom center;") is "center bottom"
PASS test("object-position: center left;") is "left center"
PASS test("object-position: center bottom;") is "center bottom"
PASS test("object-position: 100px;") is "100px center"
PASS test("object-position: 100px 100px;") is "100px 100px"
PASS test("object-position: 100px 200px;") is "100px 200px"
PASS test("object-position: -50% 0;") is "-50% 0px"
PASS test("object-position: 3em 0;") is "3em 0px"
PASS test("object-position: left 33px;") is "left 33px"
PASS test("object-position: center 33px;") is "center 33px"
PASS test("object-position: 33px center;") is "33px center"
PASS test("object-position: 33px bottom;") is "33px bottom"
PASS test("object-position: 1vh 1vw;") is "1vh 1vw"
PASS test("object-position: left 20px bottom 50%;") is "left 20px bottom 50%"
PASS test("object-position: right 50% bottom 25%;") is "right 50% bottom 25%"
PASS test("object-position: bottom 20px right 12px;") is "right 12px bottom 20px"
PASS test("object-position: 100px 100px 100px;") is ""
PASS test("object-position: 100px 100px 200px 200px;") is ""
PASS test("object-position: center bottom 25%;") is ""
PASS test("object-position: top left center;") is ""
PASS test("object-position: top top;") is ""
PASS test("object-position: top bottom;") is ""
PASS test("object-position: 33px left;") is ""
PASS test("object-position: top 33px;") is ""
PASS test("object-position: inherit inherit;") is ""
PASS test("object-position: initial initial;") is ""
PASS test("object-position: -webkit-fill-available;") is ""
PASS test("object-position: min-content;") is ""
PASS test("object-position: intrinsic;") is ""
PASS test("object-position: auto;") is ""
PASS test("object-position: none;") is ""
PASS test("object-position: fill;") is ""
PASS successfullyParsed is true

TEST COMPLETE
